requirements.txt
streamlit>=1.38.0
google-generativeai
elevenlabs
python-dotenv
